refactor(base): give more appriate names to base modules
