+++ components/base/index.js
@@ -3,5 +3,5 @@ export * from './base-plugin.js'
 export * from './custom-element.js'
 export * from './custom-link.js'
 export * from './custom-event.js'
-export * from './event-dispatcher.js'
-export * from './input-dispatcher.js'
+export * from './dispatch-event-mixin.js'
+export * from './dispatch-focus-mixin.js'
